fix: delete revire record only when they are available

+++ src/lib/utils/project.ts
@@ -99,9 +99,15 @@ export async function updateVideoDetails(videoId: number, details: { title: stri
                     status: "PENDING"
                 }
             })
-            await prisma.review.delete({
+            const existingReview = await prisma.review.findUnique({
                 where: { projectId: projectStatus.id }
-            })
+            });
+
+            if (existingReview) {
+                await prisma.review.delete({
+                    where: { projectId: projectStatus.id }
+                });
+            }
             return { updatedVideo, projectStatus };
 
         });
